backend/requirements.txt
aiohttp==3.9.1
aioredis==2.0.1
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
asyncpg==0.29.0
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
black==23.11.0
cachetools==5.3.2
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coverage==7.10.7
cryptography==46.0.2
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
execnet==2.1.1
factory-boy==3.3.0
Faker==20.1.0
fastapi==0.104.1
filelock==3.20.0
flake8==6.1.0
freezegun==1.4.0
frozenlist==1.8.0
GeoAlchemy2==0.14.3
geographiclib==2.1
geopy==2.4.1
google-cloud-texttospeech==2.14.1
greenlet==3.2.4
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
identify==2.6.15
idna==3.11
iniconfig==2.1.0
jiter==0.11.0
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
multidict==6.7.0
mypy==1.7.1
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==1.26.2
openai==2.3.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pgvector==0.2.4
platformdirs==4.5.0
pluggy==1.6.0
pre-commit==3.5.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
propcache==0.4.1
psutil==5.9.6
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.6.1
pycodestyle==2.11.1
pycparser==2.23
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pyflakes==3.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-benchmark==4.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0
pytest-xdist==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.3
redis==5.0.1
requests==2.32.5
rsa==4.9.1
ruff==0.1.7
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.23
starlette==0.27.0
structlog==23.2.0
tenacity==8.2.3
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
uvicorn[standard]==0.24.0
gunicorn==21.2.0
virtualenv==20.35.3
watchfiles==1.1.0
websockets==15.0.1
wikipedia==1.4.0
yarl==1.22.0
